update code of sssp

+++ lonestar/analytics/distributed/sssp/sssp_pull.cpp
@@ -62,7 +62,7 @@ static cll::opt<Exec> execution(
 /* Graph structure declarations + other initialization */
 /******************************************************************************/
 
-const uint32_t infinity = std::numeric_limits<uint32_t>::max() / 4;
+const double infinity = std::numeric_limits<double>::max() / 4;
 
 struct NodeData {
   double dist_current;
@@ -82,11 +82,11 @@ std::unique_ptr<galois::graphs::GluonSubstrate<Graph>> syncSubstrate;
 /******************************************************************************/
 
 struct InitializeGraph {
-  const uint32_t& local_infinity;
+  const double& local_infinity;
   cll::opt<uint64_t>& local_src_node;
   Graph* graph;
 
-  InitializeGraph(cll::opt<uint64_t>& _src_node, const uint32_t& _infinity,
+  InitializeGraph(cll::opt<uint64_t>& _src_node, const double& _infinity,
                   Graph* _graph)
       : local_infinity(_infinity), local_src_node(_src_node), graph(_graph) {}
 
@@ -199,23 +199,23 @@ struct SSSP {
 
 /* Prints total number of nodes visited + max distance */
 struct SSSPSanityCheck {
-  const uint32_t& local_infinity;
+  const double& local_infinity;
   Graph* graph;
 
   galois::DGAccumulator<uint64_t>& DGAccumulator_sum;
-  galois::DGReduceMax<uint32_t>& DGMax;
-  galois::DGAccumulator<uint64_t>& dg_avg;
+  galois::DGReduceMax<double>& DGMax;
+  galois::DGAccumulator<double>& dg_avg;
 
-  SSSPSanityCheck(const uint32_t& _infinity, Graph* _graph,
+  SSSPSanityCheck(const double& _infinity, Graph* _graph,
                   galois::DGAccumulator<uint64_t>& dgas,
-                  galois::DGReduceMax<uint32_t>& dgm,
-                  galois::DGAccumulator<uint64_t>& _dg_avg)
+                  galois::DGReduceMax<double>& dgm,
+                  galois::DGAccumulator<double>& _dg_avg)
       : local_infinity(_infinity), graph(_graph), DGAccumulator_sum(dgas),
         DGMax(dgm), dg_avg(_dg_avg) {}
 
   void static go(Graph& _graph, galois::DGAccumulator<uint64_t>& dgas,
-                 galois::DGReduceMax<uint32_t>& dgm,
-                 galois::DGAccumulator<uint64_t>& dgag) {
+                 galois::DGReduceMax<double>& dgm,
+                 galois::DGAccumulator<double>& dgag) {
     dgas.reset();
     dgm.reset();
     dgag.reset();
@@ -240,7 +240,7 @@ struct SSSPSanityCheck {
     }
 
     uint64_t num_visited  = dgas.reduce();
-    uint32_t max_distance = dgm.reduce();
+    double max_distance = dgm.reduce();
 
     float visit_average = ((float)dgag.reduce()) / num_visited;
 
@@ -270,8 +270,8 @@ struct SSSPSanityCheck {
 /* Make results */
 /******************************************************************************/
 
-std::vector<uint32_t> makeResultsCPU(std::unique_ptr<Graph>& hg) {
-  std::vector<uint32_t> values;
+std::vector<double> makeResultsCPU(std::unique_ptr<Graph>& hg) {
+  std::vector<double> values;
 
   values.reserve(hg->numMasters());
   for (auto node : hg->masterNodesRange()) {
@@ -282,8 +282,8 @@ std::vector<uint32_t> makeResultsCPU(std::unique_ptr<Graph>& hg) {
 }
 
 #ifdef GALOIS_ENABLE_GPU
-std::vector<uint32_t> makeResultsGPU(std::unique_ptr<Graph>& hg) {
-  std::vector<uint32_t> values;
+std::vector<double> makeResultsGPU(std::unique_ptr<Graph>& hg) {
+  std::vector<double> values;
 
   values.reserve(hg->numMasters());
   for (auto node : hg->masterNodesRange()) {
@@ -293,12 +293,12 @@ std::vector<uint32_t> makeResultsGPU(std::unique_ptr<Graph>& hg) {
   return values;
 }
 #else
-std::vector<uint32_t> makeResultsGPU(std::unique_ptr<Graph>& /*unused*/) {
+std::vector<double> makeResultsGPU(std::unique_ptr<Graph>& /*unused*/) {
   abort();
 }
 #endif
 
-std::vector<uint32_t> makeResults(std::unique_ptr<Graph>& hg) {
+std::vector<double> makeResults(std::unique_ptr<Graph>& hg) {
   switch (personality) {
   case CPU:
     return makeResultsCPU(hg);
@@ -354,8 +354,8 @@ int main(int argc, char** argv) {
 
   // accumulators for use in operators
   galois::DGAccumulator<uint64_t> DGAccumulator_sum;
-  galois::DGAccumulator<uint64_t> dg_avge;
-  galois::DGReduceMax<uint32_t> m;
+  galois::DGAccumulator<double> dg_avge;
+  galois::DGReduceMax<double> m;
 
   for (auto run = 0; run < numRuns; ++run) {
     galois::gPrint("[", net.ID, "] SSSP::go run ", run, " called\n");
@@ -393,7 +393,7 @@ int main(int argc, char** argv) {
   StatTimer_total.stop();
 
   if (output) {
-    std::vector<uint32_t> results = makeResults(hg);
+    std::vector<double> results = makeResults(hg);
     auto globalIDs                = hg->getMasterGlobalIDs();
     assert(results.size() == globalIDs.size());
